fix: 修复 AI 识别菜谱份数始终为双人餐的问题
问题链路：
- 移动端 createRecipe() 硬编码 servings=2，忽略 AI 识别结果
- applyRecognition() 遗漏了 servings 字段的映射
- 后端 AI prompt 缺少份数判断规则，DeepSeek 默认返回 2

修复：
- RecipeFormState 新增 servings 字段，默认值 1
- applyRecognition() 映射 AI 返回的 servings 到表单
- createRecipe() 从表单读取 servings 而非硬编码
- 后端 prompt 增加份数判断规则（从菜名提取线索）
- 后端默认 servings fallback 从 2 改为 1
- 修复 FavoriteWithTitle.servings 类型 i32 → i16

+++ backend/features/social/src/model.rs
@@ -21,7 +21,7 @@ pub struct FavoriteWithTitle {
     pub cover_image: Option<String>,
     pub difficulty: Option<i16>,
     pub cooking_time: Option<i32>,
-    pub servings: Option<i32>,
+    pub servings: Option<i16>,
 }
 
 #[derive(Debug, FromRow, Serialize, utoipa::ToSchema)]
